apps/producer/requirements.txt
openmeteo-requests
requests-cache
retry-requests
python-dotenv
pika
schedule
numpy
pandas
openpyxl
xlsxwriter
openai
